Reduce code size on latest Rust
Workaround for https://github.com/rust-lang/rust/issues/74947 which made each indexing location include itself as a string in the final binary, resulting in a significant code bloat in case of hqx which generates a lot of indexing via macros.

Stats for a simple Wasm that reexports all 3 variants of HQX:

initial size:                         462076 bytes
Dst wrapper with unreachable!:        250225 bytes
Dst wrapper with debug_unreachable!:  190312 bytes

I'm open to removing `debug_unreachable!` in favour of just `unreachable`, but I was assuming that indexing code is well-tested and the extra difference seemed nice to have.